+++ resolve/vector/Vector.hpp
@@ -35,10 +35,12 @@ namespace ReSolve
       Vector(index_type n, index_type k);
       ~Vector();
 
-      int        copyDataFrom(const real_type* data, memory::MemorySpace memspaceIn, memory::MemorySpace memspaceOut);
-      int        copyDataFrom(Vector* v, memory::MemorySpace memspaceIn, memory::MemorySpace memspaceOut);
-      real_type* getData(memory::MemorySpace memspace);
-      real_type* getData(index_type i, memory::MemorySpace memspace);
+      int              copyDataFrom(const real_type* source, memory::MemorySpace memspaceIn, memory::MemorySpace memspaceOut);
+      int              copyDataFrom(Vector* source, memory::MemorySpace memspaceIn, memory::MemorySpace memspaceOut);
+      real_type*       getData(memory::MemorySpace memspace);
+      real_type*       getData(index_type i, memory::MemorySpace memspace);
+      const real_type* getData(memory::MemorySpace memspace) const;
+      const real_type* getData(index_type i, memory::MemorySpace memspace) const;
 
       index_type getCapacity() const;
       index_type getSize() const;